Refactor dedicated_time_protocol
All ranging code is isolated as it's own module.

app_timer will be phased out and the exclusive use of delayed tx/rx
will be used instead. This is to ensure that all tx/rx occurs at
precise moments in system time. app_timer from nRF SDK is too
unstable and does not provide the time resolution that is comparable
to the 15.6ps provided by dwm1001 system timer.